% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{alluvialServer}
\alias{alluvialServer}
\title{Alluvial plot server function}
\usage{
alluvialServer(id, obj, res_obj)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive containing GeneTonic object}

\item{res_obj}{reactive, dataframe containing enrichment results}
}
\description{
Alluvial plot server function
}
