% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{check_config}
\alias{check_config}
\title{validate object creation yaml file}
\usage{
check_config(
  config,
  rules = system.file("extdata", "create.yaml", package = "dashboard_app")
)
}
\arguments{
\item{config}{named list from reading object creation yaml}

\item{rules}{yaml file containing rules for object creation}
}
\description{
\itemize{
\item check if yaml contains required & optional keys
\item check if yaml blocks contain the allowed keys
\item check if de_analysis 'counts' match counts 'key'
\item check if paths exist
}
}
