% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{check_res_columns}
\alias{check_res_columns}
\title{check column names of DE analysis results}
\usage{
check_res_columns(res, config, id2symbol)
}
\arguments{
\item{res}{data.frame with DE analysis results}

\item{config}{named list from reading object creation yaml}

\item{id2symbol}{data.frame with gene id -> symbol mapping}
}
\description{
\itemize{
\item Required columns are c('padj', 'log2FoldChange', 'symbol')
}
}
