% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{create_object}
\alias{create_object}
\title{Create object to be used with dashboard app}
\usage{
create_object(config)
}
\arguments{
\item{config}{path to object creation config yaml}
}
\description{
Create object to be used with dashboard app
}
