% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-plot.R
\name{degmod}
\alias{degmod}
\alias{patternPlotUI}
\alias{patternPlotServer}
\title{Pattern plot module}
\usage{
patternPlotUI(id, panel, tab)

patternPlotServer(id, obj, coldata, plot_args, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{tab}{string, if 'plot' show plot settings, if 'table' show table settings;
if 'both', show settings for both.}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive containing 'gene_scratchpad' (genes selected in scratchpad) &
'upset_data' (list containing data from upset plot module)}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with module UI
server invisibly returns NULL (used for side effects)
}
\description{
Module UI & server to generate pattern plots.
}
