% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{distillPlotServer}
\alias{distillPlotServer}
\title{Distill enrichment plot server function}
\usage{
distillPlotServer(id, obj, args, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive containing 'distilled' enrichment results}

\item{args}{reactive, list with plot arguments, 'numcat' (number of categories to plot)}

\item{config}{reactive list with config settings}
}
\description{
Distill enrichment plot server function
}
