% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{downloadButtonServer}
\alias{downloadButtonServer}
\title{Download button module server}
\usage{
downloadButtonServer(id, outplot, plot_type)
}
\arguments{
\item{id}{Input id}

\item{outplot}{reactive plot handle}

\item{plot_type}{reactive/static value used for output filename}
}
\description{
Server for download button module
}
