% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dummy_genetonic}
\alias{dummy_genetonic}
\title{Make dummy GeneTonic object}
\usage{
dummy_genetonic(eres)
}
\arguments{
\item{eres}{enrichResult object}
}
\value{
GeneTonic object
}
\description{
Make dummy GeneTonic object
}
