% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enrich_to_genetonic}
\alias{enrich_to_genetonic}
\title{Convert enrichResult to GeneTonic object}
\usage{
enrich_to_genetonic(enrich, res)
}
\arguments{
\item{enrich}{enrichResult object}

\item{res}{data frame with DE analysis results}
}
\value{
GeneTonic object
}
\description{
This function takes an enrichResult object and
DE analysis results and creates a GeneTonic object.
}
\examples{
# get enrich & res objects
data(res_dex, package="carnation")
data(enrich_bp_dex, package="carnation")

# convert to GeneTonic object
gt <- enrich_to_genetonic(eres, res)


}
