% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{enrichmapServer}
\alias{enrichmapServer}
\title{Enrichment map plot server function}
\usage{
enrichmapServer(id, obj, res_obj, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing GeneTonic object}

\item{res_obj}{reactive, dataframe containing enrichment results}

\item{config}{reactive list with config settings}
}
\description{
Enrichment map plot server function
}
