% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{enrichServer}
\alias{enrichServer}
\title{Functional enrichment module server function}
\usage{
enrichServer(id, obj, upset_table, gene_scratchpad, reset_genes, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{upset_table}{reactive, data from upset plot module}

\item{gene_scratchpad}{reactive, genes selected in gene scratchpad}

\item{reset_genes}{reactive to reset genes in scratchpad}

\item{config}{reactive list with config settings}
}
\description{
Functional enrichment module server function
}
