% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{enrichUI}
\alias{enrichUI}
\title{Functional enrichment module UI}
\usage{
enrichUI(id, panel, tab = "none")
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{tab}{string, if 'table' show table settings, if 'plots' show plot settings;
if 'compare_results', show comparison settings.}
}
\description{
Functional enrichment module UI
}
