% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional-enrichment.R
\name{funenrichmod}
\alias{funenrichmod}
\alias{enrichUI}
\alias{enrichServer}
\title{Functional enrichment module}
\usage{
enrichUI(id, panel, tab = "none")

enrichServer(id, obj, upset_table, gene_scratchpad, reset_genes, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{tab}{string, if 'table' show table settings, if 'plots' show plot settings;
if 'compare_results', show comparison settings.}

\item{obj}{reactiveValues object containing carnation object}

\item{upset_table}{reactive, data from upset plot module}

\item{gene_scratchpad}{reactive, genes selected in gene scratchpad}

\item{reset_genes}{reactive to reset genes in scratchpad}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with plot UI
server returns reactive with gene selected from functional enrichment tables.
}
\description{
UI & module to show functional enrichment tables & plots.
}
