% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{fuzzyPlotServer}
\alias{fuzzyPlotServer}
\title{Fuzzy enrichment plot server function}
\usage{
fuzzyPlotServer(id, obj, args)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive containing fuzzy enrichment object}

\item{args}{reactive, list with plot arguments, 'numcat' (number of categories to plot)}
}
\description{
Fuzzy enrichment plot server function
}
