% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-plot.R
\name{geneplotmod}
\alias{geneplotmod}
\alias{genePlotUI}
\alias{genePlotServer}
\title{Gene plot module}
\usage{
genePlotUI(id, panel)

genePlotServer(id, obj, coldata, plot_args, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive list with 3 elements: 'gene.id' (all gene IDs) & 'gene_scratchpad'
(genes selected in scratchpad) & 'comp_all' (selected comparison)}

\item{config}{reactive list with config settings}
}
\description{
UI & server for module to create gene plot
}
\examples{
library(DESeq2)
library(shiny)

dds <- makeExampleDESeqDataSet()
rld <- varianceStabilizingTransformation(dds, blind=TRUE)

dds <- DESeq(dds)
results <- results(dds, contrast=c('condition', 'A', 'B'))

# Create reactive values to simulate app state
obj <- reactiveValues(
  dds = list(main = dds),
  rld = list(main = rld),
  res = list(comp1 = results),
  all_dds = dds,
  all_rld = rld,
  dds_mapping = list(comp1 = 'main')
)

# Set up coldata structure that the module expects
coldata <- reactiveValues(
  curr = list(
    all_samples = colData(dds),
    main = colData(dds)
  )
)

plot_args <- reactive({
  list(
    gene.to.plot = c("gene1", "gene2"),
    gene.id = rownames(dds),
    comp_all = "comp1"
  )
})

config <- reactiveVal(get_config())

if(interactive()){
  shinyApp(
    ui = fluidPage(
           sidebarPanel(genePlotUI('p', 'sidebar')),
           mainPanel(genePlotUI('p', 'main'))
         ),
    server = function(input, output, session){
               genePlotServer('p', obj, coldata, plot_args, config)
             }
  )
}

}
