% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-plot.R
\name{genePlotServer}
\alias{genePlotServer}
\title{Gene plot module server function}
\usage{
genePlotServer(id, obj, coldata, plot_args)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive list with 3 elements: 'gene.id' (all gene IDs) & 'gene_scratchpad'
(genes selected in scratchpad) & 'comp_all' (selected comparison)}
}
\description{
Gene plot module server function
}
