% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-plot.R
\name{genePlotUI}
\alias{genePlotUI}
\title{Gene plot module UI}
\usage{
genePlotUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
Gene plot module UI
}
