% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{get_access_path}
\alias{get_access_path}
\title{Get path to access yaml file}
\usage{
get_access_path()
}
\value{
path to access yaml
}
\description{
This function checks for an environment variable 'CARNATION_ACCESS_YAML'
to specify directory to save access yaml. If env variable does not exist
uses home directory as save location.
}
\examples{
p <- get_access_path()

}
