% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_common_path_from_list}
\alias{get_common_path_from_list}
\title{Return common base path from list of paths}
\usage{
get_common_path_from_list(df)
}
\arguments{
\item{df}{data frame with two columns: data_area & user_group}
}
\value{
vector of base paths
}
\description{
Return common base path from list of paths
}
