% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-functions.R
\name{get_config_path}
\alias{get_config_path}
\title{Get path to local config yaml file}
\usage{
get_config_path()
}
\value{
path to local config yaml
}
\description{
This function checks for an environment variable
\code{CARNATION_CONFIG_YAML} to specify the local config yaml path.
If the variable is not set, a default path in the home directory is used.
}
\examples{
p <- get_config_path()

}
