% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_upset_table}
\alias{get_upset_table}
\title{Generate upset plot table}
\usage{
get_upset_table(gene.lists, comp_split_pattern = ";")
}
\arguments{
\item{gene.lists}{list with character vectors of gene names}

\item{comp_split_pattern}{character used to separate gene set names}
}
\value{
list with upset table elements
}
\description{
Generate upset plot table
}
\examples{
lst <- list(group1 = c(a = "gene1", b = "gene2", c = "gene3", d = "gene4"),
            group2 = c(b = "gene2", d = "gene4", e = "gene5"),
            group3 = c(d = "gene4", e = "gene5", f = "gene6"))

df <- get_upset_table(lst)
str(df)

}
