% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_y_init}
\alias{get_y_init}
\title{Get initial y-axis limits}
\usage{
get_y_init(df, y_delta, pseudocount)
}
\arguments{
\item{df}{data.frame with counts. Must have column 'count'}

\item{y_delta}{y-axis padding for visualization, must be
between 0 and 1}

\item{pseudocount}{pseudo-count to add to the data.frame}
}
\description{
Get initial y-axis limits
}
