% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapServer}
\alias{heatmapServer}
\title{Heatmap module server function}
\usage{
heatmapServer(id, obj, coldata, plot_args, gene_scratchpad, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold), 'fc.thres' (log2FC) &
'upset_data' (list containing data from upset plot module)}

\item{gene_scratchpad}{reactiveValues object containing genes selected in scratchpad which will
be labeled}

\item{config}{reactive list with config settings}
}
\description{
Heatmap module server function
}
