% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapUI}
\alias{heatmapUI}
\title{Heatmap module UI}
\usage{
heatmapUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
Heatmap module UI
}
