% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{horizonServer}
\alias{horizonServer}
\title{Horizon plot server function}
\usage{
horizonServer(id, obj)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive containing GeneTonic object}
}
\description{
Horizon plot server function
}
