% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{horizonUI}
\alias{horizonUI}
\title{Horizon plot UI}
\usage{
horizonUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
Horizon plot UI
}
