% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{in_admin_group}
\alias{in_admin_group}
\title{is user is in admin group?}
\usage{
in_admin_group(u)
}
\arguments{
\item{u}{username}
}
\value{
TRUE/FALSE to indicate if the user is part of the admin group
}
\description{
is user is in admin group?
}
\examples{
check <- in_admin_group('admin')

}
