% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_site_admin}
\alias{is_site_admin}
\title{is user an admin?}
\usage{
is_site_admin(u)
}
\arguments{
\item{u}{username}
}
\value{
boolean to indicate is user is in admin group
}
\description{
is user an admin?
}
\examples{
# check if default user is admin
yy <- is_site_admin(u='admin')

}
