% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-new-data.R
\name{loadDataUI}
\alias{loadDataUI}
\title{Load data module UI}
\usage{
loadDataUI(id)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}
}
\description{
Load data module UI
}
