% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_example_carnation_object}
\alias{make_example_carnation_object}
\title{Make example carnation object}
\usage{
make_example_carnation_object()
}
\value{
reactiveValues object containing carnation object
}
\description{
Returns example carnation object used in examples & testing
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
obj <- make_example_carnation_object()
\dontshow{\}) # examplesIf}
}
