% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_final_object}
\alias{make_final_object}
\title{Make final object for internal use by the app}
\usage{
make_final_object(obj)
}
\arguments{
\item{obj}{list object containing lists of DE analysis
results, functional enrichment objects, pattern analysis
objects & raw and normalized counts objects.}
}
\description{
This function takes an uploaded object and sanitizes
it to make sure it is suitable for internal use along
with other additions:
\itemize{
\item adds a 'dds_mapping' element that maps
dds_list keys to res_list objects.
\item if there are multiple dds_list objects, it adds
a 'all_dds' element combining all samples.
}
}
