% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-plot.R
\name{maPlotServer}
\alias{maPlotServer}
\title{MA plot module server function}
\usage{
maPlotServer(id, obj, plot_args, config)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold), 'fc.thres' (log2FC threshold)
& 'gene.to.plot' (genes selected in scratchpad)}

\item{config}{reactive list with config settings}
}
\description{
MA plot module server function
}
