% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-plot.R
\name{maPlotUI}
\alias{maPlotUI}
\title{MA plot module UI}
\usage{
maPlotUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
MA plot module UI
}
