% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataServer}
\alias{metadataServer}
\title{Metadata module server}
\usage{
metadataServer(id, obj, cols.to.drop)
}
\arguments{
\item{id}{Input id}

\item{obj}{internal app object}

\item{cols.to.drop}{columns to hide from table}
}
\description{
Server code for settings module
}
