% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataUI}
\alias{metadataUI}
\title{Metadata module ui}
\usage{
metadataUI(id, panel)
}
\arguments{
\item{id}{Input id}

\item{panel}{context for generating ui elements ('sidebar' or 'main')}
}
\description{
This generates the metadata tab that allows users to
view the metadata attached to the objects being viewed.
}
