% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-plot.R
\name{patternPlotServer}
\alias{patternPlotServer}
\title{Pattern plot module server function}
\usage{
patternPlotServer(id, obj, coldata, plot_args)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive containing 'gene_scratchpad' (genes selected in scratchpad) &
'upset_data' (list containing data from upset plot module)}
}
\description{
Pattern plot module server function
}
