% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-plot.R
\name{patternPlotUI}
\alias{patternPlotUI}
\title{Pattern plot module UI}
\usage{
patternPlotUI(id, panel, tab)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{tab}{string, if 'plot' show plot settings, if 'table' show table settings;
if 'both', show settings for both.}
}
\description{
Pattern plot module UI
}
