% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{pcaPlotServer}
\alias{pcaPlotServer}
\title{PCA plot module server function}
\usage{
pcaPlotServer(id, obj, coldata)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}
}
\description{
PCA plot module server function
}
