% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{pcaPlotUI}
\alias{pcaPlotUI}
\title{PCA plot module UI}
\usage{
pcaPlotUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
PCA plot module UI
}
