% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{radarServer}
\alias{radarServer}
\title{Radar plot server function}
\usage{
radarServer(id, obj, type = "")
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive containing GeneTonic object}

\item{type}{string, if 'comp' then show the comparison view}
}
\description{
Radar plot server function
}
