% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{radarUI}
\alias{radarUI}
\title{Radar plot UI}
\usage{
radarUI(id, panel, type = "")
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{type}{string, if 'comp' then show the comparison view}
}
\description{
Radar plot UI
}
