% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnation-data.R
\docType{data}
\name{res_cell}
\alias{res_cell}
\title{A \code{DESeqResults} object testing the difference between
two cell lines of smooth muscle cells}
\description{
A \code{DESeqResults} object, generated in the \code{DESeq2} framework
}
\details{
This \code{DESeqResults} object on the data from the \code{airway} package
has been created comparing two smooth muscle cell lines,
accounting for the effect of dexamethasone treatment.

Details on how this object has been created are included in the
\code{create_carnation_data.R} script, included in the \code{scripts} folder of the
\code{Carnation} package.
}
\references{
Himes BE, Jiang X, Wagner P, Hu R, Wang Q, Klanderman B, Whitaker
RM, Duan Q, Lasky-Su J, Nikolos C, Jester W, Johnson M, Panettieri R Jr,
Tantisira KG, Weiss ST, Lu Q. “RNA-Seq Transcriptome Profiling Identifies
CRISPLD2 as a Glucocorticoid Responsive Gene that Modulates Cytokine Function
in Airway Smooth Muscle Cells.” PLoS One. 2014 Jun 13;9(6):e99625. PMID:
24926665. GEO: GSE52778
}
