% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_carnation}
\alias{run_carnation}
\title{Carnation}
\usage{
run_carnation(
  config = NULL,
  credentials = NULL,
  passphrase = NULL,
  enable_admin = TRUE,
  ...
)
}
\arguments{
\item{config}{list of config options. If NULL, uses internal config.}

\item{credentials}{path to encrypted sqlite db with user credentials.}

\item{passphrase}{passphrase for credentials db.}

\item{enable_admin}{if TRUE, admin view is shown. Note, this is only available
if credentials have sqlite backend.}

\item{...}{parameters passed to shinyApp() call}
}
\description{
Interactive shiny dashboard for exploring RNA-Seq analysis.
}
