% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{savemod}
\alias{savemod}
\alias{saveUI}
\alias{saveServer}
\title{Save object module UI}
\usage{
saveUI(id)

saveServer(id, original, current, coldata, pattern, username, config)
}
\arguments{
\item{id}{Module id}

\item{original}{original carnation object}

\item{current}{current carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{pattern}{regex pattern for finding carnation data}

\item{username}{user name}

\item{config}{reactive list with config settings}
}
\value{
UI returns actionButton
Server returns reactive with trigger to refresh the app
}
\description{
Module UI & server to save carnation object.
}
