% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{saveServer}
\alias{saveServer}
\title{Save object module server function}
\usage{
saveServer(id, original, current, coldata, pattern)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{original}{original carnation object}

\item{current}{current carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{pattern}{regex pattern for finding carnation data}
}
\description{
Save object module server function
}
