% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{saveUI}
\alias{saveUI}
\title{Save object module UI}
\usage{
saveUI(id)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}
}
\description{
Save object module UI
}
