% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsServer}
\alias{settingsServer}
\title{Settings module server}
\usage{
settingsServer(id, username, depth, end_offset, assay_fun)
}
\arguments{
\item{id}{Input id}

\item{username}{user name}

\item{depth}{project name depth}

\item{end_offset}{project name end offset}

\item{assay_fun}{function to parse assay names from file path}
}
\description{
Server code for settings module
}
