% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{sumovPlotServer}
\alias{sumovPlotServer}
\title{Summary overview plot server function}
\usage{
sumovPlotServer(id, obj, type = "")
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing GeneTonic object}

\item{type}{string, if 'comp' then show the comparison view}
}
\description{
Summary overview plot server function
}
