% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset-plot.R
\name{upsetmod}
\alias{upsetmod}
\alias{upsetPlotUI}
\alias{upsetPlotServer}
\title{Upset plot module}
\usage{
upsetPlotUI(id, panel)

upsetPlotServer(id, obj, plot_args, gene_scratchpad, reset_genes, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{obj}{reactiveValues object containing carnation object}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold) & 'fc.thres' (log2FC)}

\item{gene_scratchpad}{reactiveValues object containing genes selected in scratchpad}

\item{reset_genes}{reactive to reset gene scratchpad selection}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with heatmap UI.
Server invisibly returns NULL (used for side effects).
}
\description{
Module UI & server to generate upset plots.
}
