% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset-plot.R
\name{upsetPlotServer}
\alias{upsetPlotServer}
\title{Upset plot module server function}
\usage{
upsetPlotServer(id, obj, plot_args, gene_scratchpad, reset_genes)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold) & 'fc.thres' (log2FC)}

\item{gene_scratchpad}{reactiveValues object containing genes selected in scratchpad}

\item{reset_genes}{reactive to reset gene scratchpad selection}
}
\description{
Upset plot module server function
}
