% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset-plot.R
\name{upsetPlotUI}
\alias{upsetPlotUI}
\title{Upset plot module UI}
\usage{
upsetPlotUI(id, panel)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module server function}

\item{panel}{string, can be 'sidebar' or 'main'}
}
\description{
Upset plot module UI
}
