% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{alluvialmod}
\alias{alluvialmod}
\alias{alluvialUI}
\alias{alluvialServer}
\title{Alluvial plot module}
\usage{
alluvialUI(id, panel)

alluvialServer(id, obj, res_obj, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{obj}{reactiveValues object containing GeneTonic object}

\item{res_obj}{reactive, dataframe containing enrichment results}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with plot UI
server invisibly returns NULL (used for side effects)
}
\description{
UI & module to generate alluvial plots.
}
\examples{
library(airway)
library(DESeq2)
library(org.Hs.eg.db)
library(shiny)

# load airway data
data('airway')

# extract counts and metadata
mat <- assay(airway)
cdata <- colData(airway)

# get symbol annotations
anno_df <- mapIds(org.Hs.eg.db,
               column='SYMBOL',
               keys=rownames(mat),
               keytype='ENSEMBL')

# analyze with DESeq2
dds <- DESeqDataSetFromMatrix(mat,
                              colData=cdata,
                              design=~cell + dex)
dds <- DESeq(dds)

# extract comparison of interest
res <- results(dds, contrast = c("dex", "trt", "untrt"))

# add gene column from rownames
res$gene <- rownames(res)

# add symbol column from annotations
res$symbol <- anno_df[rownames(res)]


# get DE genes with FDR < 0.1
de_genes <- rownames(res)[res$padj < 0.1 & !is.na(res$padj)]

\donttest{
# functional enrichment using GO BP
eres <- clusterProfiler::enrichGO(
            gene = de_genes[1:100],
            keyType = 'ENSEMBL',
            OrgDb=org.Hs.eg.db,
            ont='BP'
        )

# convert to GeneTonic object
gt <- GeneTonic::shake_enrichResult(eres)

obj <- reactive({
  list(l_gs = gt$l_gs,
       anno_df = gt$anno_df,
       label = 'comp1')
})

res_obj <- reactive({ res })

config <- reactiveVal(get_config())

# run simple shiny app with plot
if(interactive()){
  shinyApp(
    ui = fluidPage(
           sidebarPanel(alluvialUI('p', 'sidebar')),
           mainPanel(alluvialUI('p', 'main'))
         ),
    server = function(input, output, session){
               alluvialServer('p', obj, res_obj, config)
             }
  )
}

}
}
