% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnation-pkg.R
\docType{package}
\name{carnation-package}
\alias{carnation}
\alias{carnation-package}
\title{carnation}
\description{
\code{carnation} is an interactive Shiny dashboard that makes complex bulk RNA-Seq
data more accessible and intuitive, integrating all facets of bulk RNA-Seq
analysis using three modules - differential expression analysis, functional
enrichment and pattern analysis.
}
\details{
\itemize{
\item Deeply explore analysis results from complex experiments using interactive plots.
\item Easily keep track of genes of interest using the 'Gene scratchpad'.
\item Use fuzzy search to filter and search functional enrichment results.
\item Visualize complex patterns using highly customizable gene plot.
\item Manage local data in single-user mode or deploy on a server to share with
collaborators using in-built user management system.
}

Main function to run the app: \code{\link[=run_carnation]{run_carnation()}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nichd-bspc.github.io/carnation/}
  \item Report bugs at \url{https://github.com/NICHD-BSPC/carnation/issues}
}

}
\author{
\strong{Maintainer}: Apratim Mitra \email{apratim.mitra@gmail.com} (\href{https://orcid.org/0000-0003-3279-0054}{ORCID})

Authors:
\itemize{
  \item Apratim Mitra \email{apratim.mitra@gmail.com} (\href{https://orcid.org/0000-0003-3279-0054}{ORCID})
}

Other contributors:
\itemize{
  \item Matthew Tyler Menold \email{matthew.menold@gmail.com} (\href{https://orcid.org/0009-0007-4728-2470}{ORCID}) [contributor]
  \item Ryan Dale \email{ryan.dale@nih.gov} (\href{https://orcid.org/0000-0003-2664-3744}{ORCID}) [funder]
}

}
\keyword{internal}
