% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{cnetPlotServer}
\alias{cnetPlotServer}
\title{Cnetplot server function}
\usage{
cnetPlotServer(id, obj)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactive, dataframe with enrichment results}
}
\description{
Cnetplot server function
}
