% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-plots.R
\name{distillmod}
\alias{distillmod}
\alias{distillPlotUI}
\alias{distillPlotServer}
\title{Distilled enrichment map module}
\usage{
distillPlotUI(id, panel)

distillPlotServer(id, obj, args, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{obj}{reactive containing 'distilled' enrichment results}

\item{args}{reactive, list with plot arguments, 'numcat' (number of categories to plot)}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with plot UI
server returns reactive with number of plotted terms
}
\description{
UI & module to generate distill enrichment map plots.
}
\examples{
\donttest{
library(airway)
library(DESeq2)
library(org.Hs.eg.db)
library(GeneTonic)
library(shiny)

# load airway data
data('airway')

# extract counts and metadata
mat <- assay(airway)
cdata <- colData(airway)

# get symbol annotations
anno_df <- mapIds(org.Hs.eg.db,
               column='SYMBOL',
               keys=rownames(mat),
               keytype='ENSEMBL')

# analyze with DESeq2
dds <- DESeqDataSetFromMatrix(mat,
                              colData=cdata,
                              design=~cell + dex)
dds <- DESeq(dds)

# extract comparison of interest
res <- results(dds, contrast = c("dex", "trt", "untrt"))

# add gene column from rownames
res$gene <- rownames(res)

# add symbol column from annotations
res$symbol <- anno_df[rownames(res)]


# get DE genes with FDR < 0.1
de_genes <- rownames(res)[res$padj < 0.1 & !is.na(res$padj)]

# functional enrichment using GO BP
eres <- clusterProfiler::enrichGO(
            gene = de_genes[1:100],
            keyType = 'ENSEMBL',
            OrgDb=org.Hs.eg.db,
            ont='BP'
        )

# convert to GeneTonic object
gt <- GeneTonic::shake_enrichResult(eres)

# get distilled results
df <- distill_enrichment(
        gt,
        res,
        ganno_df,
        n_gs = 10,
        cluster_fun = "cluster_markov"
      )

# number of plotted terms
args <- reactive({ list(numcat=10) })

config <- reactiveVal(get_config())

# run simple shiny app with plot
if(interactive()){
  shinyApp(
    ui = fluidPage(
           sidebarPanel(distillPlotUI('p', 'sidebar')),
           mainPanel(distillPlotUI('p', 'main'))
         ),
    server = function(input, output, session){
               numcat <- observe({
                 distillPlotServer('p',
                                   reactive({ df }),
                                   args,
                                   config)
               })
             }
  )
}

}
}
