% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gs_radar}
\alias{gs_radar}
\title{Radar plot}
\usage{
gs_radar(
  res_enrich,
  res_enrich2 = NULL,
  label1 = "scenario 1",
  label2 = "scenario 2",
  n_gs = 20,
  p_value_column = "gs_pvalue"
)
}
\arguments{
\item{res_enrich}{GeneTonic object for comparison 1}

\item{res_enrich2}{GeneTonic object for comparison 2 (default = NULL)}

\item{label1}{label for comparison 1}

\item{label2}{label for comparison 2}

\item{n_gs}{number of gene sets (default = 20)}

\item{p_value_column}{column to use as p-value (default = 'gs_pvalue')}
}
\value{
ggplot handle
}
\description{
This is a copy of gs_radar from GeneTonic where the labels of
gene sets are converted to parameters
}
\examples{

# get enrichResult object
data(enrich_bp_dex, package='carnation')

# convert to GeneTonic object
gt <- GeneTonic::shake_enrichResult(eres_dex)

# make radar plot
p <- gs_radar(gt)

}
