% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapmod}
\alias{heatmapmod}
\alias{heatmapUI}
\alias{heatmapServer}
\title{Heatmap module}
\usage{
heatmapUI(id, panel)

heatmapServer(id, obj, coldata, plot_args, gene_scratchpad, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main'}

\item{obj}{reactiveValues object containing carnation object}

\item{coldata}{reactiveValues object containing object metadata}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold), 'fc.thres' (log2FC) &
'upset_data' (list containing data from upset plot module)}

\item{gene_scratchpad}{reactiveValues object containing genes selected in scratchpad which will
be labeled}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with heatmap UI.
Server invisibly returns NULL (used for side effects).
}
\description{
Module UI & server to generate heatmap.
}
