% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-new-data.R
\name{loadmod}
\alias{loadmod}
\alias{loadDataUI}
\alias{loadDataServer}
\title{Load data module UI}
\usage{
loadDataUI(id)

loadDataServer(id, username, config, rds = NULL)
}
\arguments{
\item{id}{Module id}

\item{username}{user name}

\item{config}{reactive list with config settings}

\item{rds}{Object to be edited}
}
\value{
UI returns tagList with module UI
Server returns reactive with app reload trigger
}
\description{
Module UI & server to load new data
}
