% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotMA.label_ly}
\alias{plotMA.label_ly}
\title{Create an interactive labeled MA plot}
\usage{
plotMA.label_ly(
  res,
  fdr.thres = 0.01,
  fc.thres = 0,
  fc.lim = NULL,
  lab.genes = NULL,
  tolower.cols = c("SYMBOL", "ALIAS")
)
}
\arguments{
\item{res}{data.frame with DE analysis results. Must contain
"padj" & "log2FoldChange" columns}

\item{fdr.thres}{False discovery rate (FDR) threshold}

\item{fc.thres}{log2FoldChange threshold}

\item{fc.lim}{y-axis limits}

\item{lab.genes}{genes to label on MA plot}

\item{tolower.cols}{column names that will be converted to
lower case}
}
\value{
plotly handle
}
\description{
This function creates an MA plot from a data.frame
containing DE analysis results using plot_ly
}
\examples{
# get DE results
data(res_dex, package='carnation')

# generate interactive MA plot
plotMA.label_ly(res_dex, lab.genes = c("SPARCL1", "CACNB2"))

}
