% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotVolcano.label_ly}
\alias{plotVolcano.label_ly}
\title{Create an interactive labeled Volcano plot}
\usage{
plotVolcano.label_ly(
  res,
  fdr.thres = 0.1,
  fc.thres = 0,
  y.axis.lim = NULL,
  fc.lim = c(-4, 4),
  lab.genes = NULL,
  tolower.cols = c("SYMBOL", "ALIAS"),
  color_by = c("baseMean", "significance"),
  colorscale = NULL,
  alpha = 0.6,
  pvalue_padj_switch = c("pvalue", "padj")
)
}
\arguments{
\item{res}{data.frame with DE analysis results. Must contain
"pvalue", "padj", "log2FoldChange", & "baseMean" columns}

\item{fdr.thres}{False discovery rate (FDR) threshold}

\item{fc.thres}{log2FoldChange threshold}

\item{y.axis.lim}{y-axis limits for negative log10 p-values}

\item{fc.lim}{x-axis limits for log2FoldChange}

\item{lab.genes}{genes to label on volcano plot}

\item{tolower.cols}{column names that will be converted to
lower case}

\item{color_by}{variable to color points by ('baseMean' or 'significance')}

\item{colorscale}{ggplot-compatible continuous colorscale name
(viridis option), only used when color_by='baseMean'}

\item{alpha}{point opacity between 0 and 1}

\item{pvalue_padj_switch}{Switches between having the y-axis display raw or adjusted p-values.
Significance is always determined by adjusted values.}
}
\value{
plotly handle
}
\description{
This function creates an Volcano plot from a data.frame
containing DE analysis results using plot_ly
}
\examples{
# make mock results df
n_genes <- 100
res <- data.frame(
         baseMean = runif(n_genes, 10, 1000),
         log2FoldChange = rnorm(n_genes, 0, 2),
         lfcSE = runif(n_genes, 0.1, 0.5),
         stat = rnorm(n_genes, 0, 3),
         pvalue = runif(n_genes, 0, 1),
         padj = runif(n_genes, 0, 1),
         symbol = paste0("GENE", 1:n_genes),
         row.names = paste0("gene", 1:n_genes)
       )

plotVolcano.label_ly(res, lab.genes = c("gene1", "gene2"))

}
