% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_carnation}
\alias{run_carnation}
\title{Carnation}
\usage{
run_carnation(username = NULL, config = NULL, ...)
}
\arguments{
\item{username}{user name. When NULL (default), carnation runs in single-user mode.}

\item{config}{list of config options. If NULL, uses internal config.}

\item{...}{parameters passed to shinyApp() call}
}
\description{
Interactive shiny dashboard for exploring RNA-Seq analysis.
}
