% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plot.R
\name{scattermod}
\alias{scattermod}
\alias{scatterPlotUI}
\alias{scatterPlotServer}
\title{Scatterplot module}
\usage{
scatterPlotUI(id, panel)

scatterPlotServer(id, obj, plot_args, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{string, can be 'sidebar' or 'main' passed to UI}

\item{obj}{reactiveValues object containing carnation object passed to server}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold), 'fc.thres' (log2FC) &
'gene.to.plot' (genes to be labeled) passed to server}

\item{config}{reactive list with config settings passed to server}
}
\value{
UI returns tagList with scatter plot UI.
Server invisibly returns NULL (used for side effects).
}
\description{
Module UI + server for generating scatter plots.
}
\examples{
library(DESeq2)
library(shiny)

dds <- makeExampleDESeqDataSet()
rld <- varianceStabilizingTransformation(dds, blind=TRUE)

dds <- DESeq(dds)
results <- results(dds, contrast=c('condition', 'A', 'B'))

# Create reactive values to simulate app state
obj <- reactiveValues(
  dds = list(main = dds),
  rld = list(main = rld),
  res = list(comp1 = results),
  all_dds = dds,
  all_rld = rld,
  dds_mapping = list(comp1 = 'main')
)

plot_args <- reactive({
  list(
    fdr.thres=0.1,
    fc.thres=0,
    gene.to.plot=c('gene1', 'gene2')
  )
})

config <- reactiveVal(get_config())

if(interactive()){
  shinyApp(
    ui = fluidPage(
           sidebarPanel(scatterPlotUI('p', 'sidebar')),
           mainPanel(scatterPlotUI('p', 'sidebar'))
         ),
    server = function(input, output, session){
               scatterPlotServer('p', obj, plot_args, config)
             }
  )
}

}
