% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plot.R
\name{scatterPlotServer}
\alias{scatterPlotServer}
\title{Scatterplot module server function}
\usage{
scatterPlotServer(id, obj, plot_args)
}
\arguments{
\item{id}{ID string used to match the ID used to call the module UI function}

\item{obj}{reactiveValues object containing carnation object}

\item{plot_args}{reactive containing 'fdr.thres' (padj threshold), 'fc.thres' (log2FC) &
'gene.to.plot' (genes to be labeled)}
}
\description{
Scatterplot module server function
}
