% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsmod}
\alias{settingsmod}
\alias{settingsUI}
\alias{settingsServer}
\title{Settings module}
\usage{
settingsUI(id, panel, username)

settingsServer(id, details, depth, end_offset, assay_fun, config)
}
\arguments{
\item{id}{Module id}

\item{panel}{context for generating ui elements ('sidebar' or 'main')}

\item{username}{user name}

\item{details}{reactive list with user name & app location details}

\item{depth}{project name depth}

\item{end_offset}{project name end offset}

\item{assay_fun}{function to parse assay names from file path}

\item{config}{reactive list with config settings}
}
\value{
UI returns tagList with module UI
Server returns reactive with list containing user access details
}
\description{
Module UI & server for user access details interface.

Server code for settings module
}
